be smarter - have app handler handle it's own changes, file watcher should only emit what file changed
